s4:torture:smb2: fix the durable-open.delete_on_close1 test
This test used to flap because there was an invalid check of
the file-id. This might or might not be the same after a
new open. Hence the flapping.

The new version simply opens a file as durable handle with
delete_on_close set, writes a byte and closes the file.
Then on a new connect, the file is opened again, and it is checked
that the file has been created (again) and that it is empty.