s3fs-smbd: Move housekeeping to the background process.
If you add 200 printers using lpadmin. Then you wait for the printcap
cache to expire. As soon as this expires we notify all deamons that they
should reload the printers. This mean we need to create the default
registry keys for each printer. If you do e.g. a 'smbclient -L' during
that time you will get a lot of timeouts.

This lets the housekeeping function of the printcap cache do the task of
creating the default registry keys in background queue process. When it is done
with the task it will tell all smbd childs to reload the printers and the 200
printers appear.

(cherry picked from commit aa6a7a97803b1106d2200c889a2260f81059c450)

Signed-off-by: Andreas Schneider <asn@samba.org>

Fix bug #9231 - NT_STATUS_IO_TIMEOUT during slow import of printers into
registry.