s3:smbd/durable: update the low level write_time before disconnecting the file.
If we close the low level FD, we should better update the write_time
if needed.

It's not unlikely that the client doesn't reclaim the durable handle,
in that case we may not close the after the durable timeout.

In such a case we should make sure that we at least update the write time
on disconnect, this makes sure backup applications notice that
the file was changed.

metze

Signed-off-by: Stefan Metzmacher <metze@samba.org>
Reviewed-by: Michael Adam <obnox@samba.org>
(cherry picked from commit 137376391d5642c0a18ff70a5184dac4b402fa42)

Signed-off-by: Stefan Metzmacher <metze@samba.org>