build(waf): Fix the --with-cluster build
The problem with the cluster/ctdb configure checks in Samba
currently is, that the headers need to include <tdb.h>.
If there are no system tdb headers, configure fails to find
tdb headers at this stage. Since the include is only required
for some defines (TDB_DATA), the workaround is to temporarily
add the included tdb copy's include path to the include search
path in the configure test. The ctdb we run against will most
likely have been compiled against a different version of TDB
anyways...

To properly fix this, we might need to change ctdb to rely
on an external tdb library. Or to incorporate ctdb into samba
as a component that uses the same shipped tdb version.

Signed-off-by: Michael Adam <obnox@samba.org>
Signed-off-by: Volker Lendecke <vl@samba.org>
Signed-off-by: Stefan Metzmacher <metze@samba.org>

Autobuild-User(master): Michael Adam <obnox@samba.org>
Autobuild-Date(master): Tue Oct 23 18:29:26 CEST 2012 on sn-devel-104