BUG 9326: Fix net ads join message for the dns domain.
We don't get a realm back from the server which is useable as a realm on
Unix. On Unix they are case sensitive and on Windows they aren't. This
confuses uses and if we write realm they try to use it as it came back
in lowercase.

Signed-off-by: Andreas Schneider <asn@samba.org>
(cherry picked from commit 155d132434e96d69bac31dd7413ecb4d4b09b017)