build: Be consistent with the name of smbtorture binaries
This ensures that in both build systems, smbtorture3 is the source3 binary, and
smbtoture is our main smbtorture binary, built with waf.

Also included in this is the removal of bin/ndrdump4 as a special case.

This removes the last cases of binaries with different names in
each build system.

Andrew Bartlett

Reviewed-by: Andreas Schneider <asn@samba.org>
(cherry-picked from 8a2ef49dea36b103a87cf2aa63417737dd754ce6)