build: Remove --enable-smbtorture, require bin/smbtorture (from waf) for make test
This simply moves this to being a side-effect of --enable-selftest.

The flag was renamed from --enable-smbtorture4 in a recent patch.

Make test now relies on smbtorture4, and so this code to make the dependency
optional for the tests is not required any more.

Andrew Bartlett

Reviewed-by: Andreas Schneider <asn@samba.org>
(cherry-picked from 7626b5d9045c2b490b38dee7dd45ba7763740f83)