Revert "selftest: skip smb2.ioctl tests on ntvfs"
This reverts commit cf27c2fbb6e7422cb962f4c63a53515321c65a70.

The following concerns were raised with regard to the relocation of
tests from selftest/knownfail to selftest/skip.

Andrew Bartlett <abartlet@samba.org> wrote:

Unless there is some entirely undesirable side-effect, this patch throws
away valuable testing, because it removes the test of the testsuite for
failure (ie, does the test work!), and it removes the documentation of
the difference between the servers (if someone did add some this to the
ntvfs server, it would never be tested).

It also removes a test that if this unimplemented functionality is
called, that we don't crash and die.

Reviewed-by: Andrew Bartlett <abartlet@samba.org>

Autobuild-User(master): David Disseldorp <ddiss@samba.org>
Autobuild-Date(master): Mon Feb  4 15:57:29 CET 2013 on sn-devel-104